# Conflicts:
#	engine/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/engine/Main.java
#	engine/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/engine/races/webrunnerjsonraces/WEEK8_PREVIEW2.json
#	player/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/render/TempoRender.java
#	player/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/utils/pathfinding/PathfindingProblem.java
#	player/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/utils/shape/Polygon.java
#	player/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/utils/shape/PolygonAbstract.java
#	player/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/utils/shape/Rectangle.java
#	player/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/utils/shape/positionable/PositionablePolygon.java
#	player/src/test/java/fr/unice/polytech/si3/qgl/qualituriers/utils/pathfinding/AvoidObstaclesTest.java

+++ player/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/entity/boat/headquarter/headquarterutils/BoatPathFinding.java
@@ -64,7 +64,7 @@ public class BoatPathFinding {
 
 
         int destX = 0; int destY = 0;
-        double max = Config.MAX_MOVING_CASES_MARIN * Config.MAX_MOVING_CASES_MARIN;
+        double max = (double) Config.MAX_MOVING_CASES_MARIN * (double) Config.MAX_MOVING_CASES_MARIN;
 
         if ( differenceX >= 0 && differenceY >= 0 ) {
             // +X  +Y
