+++ player/src/main/java/fr/unice/polytech/si3/qgl/qualituriers/utils/shape/PolygonAbstract.java
@@ -106,5 +106,5 @@ public abstract class PolygonAbstract extends Shape {
 
     public abstract PolygonAbstract scaleFromCenter(double scale);
 
-    public abstract PolygonAbstract enlargeOf(int length);
+    public abstract PolygonAbstract enlargeOf(double length);
 }
